+++ app/components/default_components/Collapsible.tsx
@@ -2,8 +2,8 @@ import Ionicons from "@expo/vector-icons/Ionicons";
 import { PropsWithChildren, useState } from "react";
 import { StyleSheet, TouchableOpacity, useColorScheme } from "react-native";
 
-import { ThemedText } from "@/components/ThemedText";
-import { ThemedView } from "@/components/ThemedView";
+import { ThemedText } from "@/app/components/default_components/ThemedText";
+import { ThemedView } from "@/app/components/default_components//ThemedView";
 import { Colors } from "@/app/constants/Colors";
 
 export function Collapsible({
