+++ app/components/default_components/ParallaxScrollView.tsx
@@ -7,7 +7,7 @@ import Animated, {
   useScrollViewOffset,
 } from 'react-native-reanimated';
 
-import { ThemedView } from '@/components/ThemedView';
+import { ThemedView } from '@/app/components/default_components//ThemedView';
 
 const HEADER_HEIGHT = 250;
