+++ app/screens/Login.tsx
@@ -1,53 +1,21 @@
-import { ScrollView, StyleSheet, Text, View } from "react-native";
+import {ScrollView, Text, View} from 'react-native';
 
-import { SimpleLoginForm } from "../components/SimpleLoginForm";
-import { violet } from "@/app/constants/Colors";
-import { Link } from "expo-router";
+import {SimpleLoginForm} from "@/app/components/form-components/SimpleLoginForm";
+import {Link} from "expo-router";
+import {formContainerStyles} from "@/app/constants/Styles";
 
 export default function Login() {
   return (
-    <ScrollView>
-      <View style={styles.alignForm}>
-        <Text style={[styles.alignCenter, styles.bigFont]}>Login</Text>
-        <SimpleLoginForm />
-        <Link style={[styles.violetBoldText, styles.alignCenter]} href={""}>
-          Forgot password?
-        </Link>
-        <View style={[{ flexDirection: "row" }, styles.alignCenter]}>
-          <Text style={styles.simpleText}>Don't have an account yet? </Text>
-          <Link style={styles.link} href={""}>
-            Sign up
-          </Link>
+      <ScrollView>
+        <View style={formContainerStyles.alignForm}>
+          <Text style={[formContainerStyles.alignCenter, formContainerStyles.title]}>Login</Text>
+          <SimpleLoginForm onSubmit={(email, password) => {}}/>
+          <Link style={[formContainerStyles.accentBoldText, formContainerStyles.alignCenter]} href={""}>Forgot password?</Link>
+          <View style={[{flexDirection: "row"}, formContainerStyles.alignCenter]}>
+            <Text style={formContainerStyles.simpleText}>Don't have an account yet? </Text>
+            <Link style={formContainerStyles.link} href={""}>Sign up</Link>
+          </View>
         </View>
-      </View>
-    </ScrollView>
+      </ScrollView>
   );
 }
-
-const styles = StyleSheet.create({
-  alignForm: {
-    marginTop: "30%",
-    height: "60%",
-  },
-  alignCenter: {
-    marginLeft: "auto",
-    marginRight: "auto",
-  },
-  bigFont: {
-    fontSize: 22,
-    padding: 20,
-  },
-  violetBoldText: {
-    fontSize: 18,
-    fontWeight: "bold",
-    color: violet["100"],
-    padding: 40,
-  },
-  simpleText: {
-    fontSize: 16,
-  },
-  link: {
-    color: violet["100"],
-    fontSize: 16,
-  },
-});
